+++ test/sse.test.ts
@@ -10,7 +10,7 @@ import {
   getProjectPresence,
   getSSEClientCount,
   updateClientView,
-} from "../dist/services/sse.js";
+} from "../src/services/sse.ts";
 
 const projectId = "11111111-1111-4111-8111-111111111111";
 const otherProjectId = "33333333-3333-4333-8333-333333333333";